cleanup workspace editing service
