+++ src/vs/workbench/services/backup/common/backup.ts
@@ -6,8 +6,6 @@
 import { URI } from 'vs/base/common/uri';
 import { createDecorator } from 'vs/platform/instantiation/common/instantiation';
 import { ITextBufferFactory, ITextSnapshot } from 'vs/editor/common/model';
-import { IEnvironmentService } from 'vs/platform/environment/common/environment';
-import { joinPath, relativePath } from 'vs/base/common/resources';
 
 export const IBackupFileService = createDecorator<IBackupFileService>('backupFileService');
 
@@ -87,8 +85,4 @@ export interface IBackupFileService {
 	 * being made.
 	 */
 	discardAllWorkspaceBackups(): Promise<void>;
-}
-
-export function toBackupWorkspaceResource(backupWorkspacePath: string, environmentService: IEnvironmentService): URI {
-	return joinPath(environmentService.userRoamingDataHome, relativePath(URI.file(environmentService.userDataPath), URI.file(backupWorkspacePath))!);
 }
