+++ src/vs/workbench/services/workspaces/electron-browser/workspaceEditingService.ts
@@ -8,11 +8,12 @@ import { URI } from 'vs/base/common/uri';
 import * as nls from 'vs/nls';
 import { IWorkspaceContextService } from 'vs/platform/workspace/common/workspace';
 import { IJSONEditingService } from 'vs/workbench/services/configuration/common/jsonEditing';
-import { IWorkspacesService, isUntitledWorkspace } from 'vs/platform/workspaces/common/workspaces';
+import { IWorkspacesService, isUntitledWorkspace, IWorkspaceIdentifier } from 'vs/platform/workspaces/common/workspaces';
 import { WorkspaceService } from 'vs/workbench/services/configuration/browser/configurationService';
 import { IStorageService } from 'vs/platform/storage/common/storage';
 import { IExtensionService } from 'vs/workbench/services/extensions/common/extensions';
 import { IBackupFileService } from 'vs/workbench/services/backup/common/backup';
+import { toBackupWorkspaceResource } from 'vs/workbench/services/backup/electron-browser/backup';
 import { ICommandService } from 'vs/platform/commands/common/commands';
 import { isEqual, basename } from 'vs/base/common/resources';
 import { INotificationService, Severity } from 'vs/platform/notification/common/notification';
@@ -29,6 +30,7 @@ import { AbstractWorkspaceEditingService } from 'vs/workbench/services/workspace
 import { IElectronService } from 'vs/platform/electron/node/electron';
 import { isMacintosh, isWindows, isLinux } from 'vs/base/common/platform';
 import { mnemonicButtonLabel } from 'vs/base/common/labels';
+import { BackupFileService } from 'vs/workbench/services/backup/common/backupFileService';
 
 export class NativeWorkspaceEditingService extends AbstractWorkspaceEditingService {
 
@@ -39,9 +41,9 @@ export class NativeWorkspaceEditingService extends AbstractWorkspaceEditingServi
 		@IWorkspaceContextService contextService: WorkspaceService,
 		@IElectronService private electronService: IElectronService,
 		@IConfigurationService configurationService: IConfigurationService,
-		@IStorageService storageService: IStorageService,
-		@IExtensionService extensionService: IExtensionService,
-		@IBackupFileService backupFileService: IBackupFileService,
+		@IStorageService private storageService: IStorageService,
+		@IExtensionService private extensionService: IExtensionService,
+		@IBackupFileService private backupFileService: IBackupFileService,
 		@INotificationService notificationService: INotificationService,
 		@ICommandService commandService: ICommandService,
 		@IFileService fileService: IFileService,
@@ -54,7 +56,7 @@ export class NativeWorkspaceEditingService extends AbstractWorkspaceEditingServi
 		@ILabelService private readonly labelService: ILabelService,
 		@IHostService hostService: IHostService,
 	) {
-		super(jsonEditingService, contextService, configurationService, storageService, extensionService, backupFileService, notificationService, commandService, fileService, textFileService, workspacesService, environmentService, fileDialogService, dialogService, hostService);
+		super(jsonEditingService, contextService, configurationService, notificationService, commandService, fileService, textFileService, workspacesService, environmentService, fileDialogService, dialogService, hostService);
 
 		this.registerListeners();
 	}
@@ -164,6 +166,37 @@ export class NativeWorkspaceEditingService extends AbstractWorkspaceEditingServi
 		}
 
 		return true; // OK
+	}
+
+	async enterWorkspace(path: URI): Promise<void> {
+		const result = await this.doEnterWorkspace(path);
+		if (result) {
+
+			// Migrate storage to new workspace
+			await this.migrateStorage(result.workspace);
+
+			// Reinitialize backup service
+			this.environmentService.configuration.backupPath = result.backupPath;
+			this.environmentService.configuration.backupWorkspaceResource = result.backupPath ? toBackupWorkspaceResource(result.backupPath, this.environmentService) : undefined;
+			if (this.backupFileService instanceof BackupFileService) {
+				this.backupFileService.reinitialize();
+			}
+		}
+
+		// TODO@aeschli: workaround until restarting works
+		if (this.environmentService.configuration.remoteAuthority) {
+			this.hostService.reload();
+		}
+
+		// Restart the extension host: entering a workspace means a new location for
+		// storage and potentially a change in the workspace.rootPath property.
+		else {
+			this.extensionService.restartExtensionHost();
+		}
+	}
+
+	private migrateStorage(toWorkspace: IWorkspaceIdentifier): Promise<void> {
+		return this.storageService.migrate(toWorkspace);
 	}
 }
